% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create an RStudio project specific to UMR1283 storage architecture}
\usage{
create_project(
  path,
  analyst_name,
  working_directory,
  git_repository,
  mran = FALSE,
  targets = TRUE,
  python = FALSE,
  ...
)
}
\arguments{
\item{path}{A character string. A path to where the project is to be created.}

\item{analyst_name}{A character string. The name of the analyst in charge of that project.}

\item{working_directory}{A character string. A path to where outputs is to be generated.}

\item{git_repository}{A character string. URL to the git server/repository.}

\item{mran}{A boolean. If \code{TRUE}, uses \code{paste0("https://mran.microsoft.com/snapshot/", Sys.Date())} as the CRAN repository.
Default is \code{FALSE}.}

\item{targets}{A boolean. If \code{TRUE}, uses \code{use_targets()} to create directory tree for use with \code{targets}.
Default is \code{TRUE}.}

\item{python}{A boolean. If \code{TRUE}, uses \code{use_python()} to create \code{renv} directory tree for use with python.
Default is \code{FALSE}.}

\item{...}{not used}
}
\description{
Create an RStudio project specific to UMR1283 storage architecture
}
\examples{

if (interactive()) {
  create_project(
    path = "/disks/PROJECT/_test_project",
    analyst_name = "Analyst Name",
    working_directory = "/disks/DATATMP",
    git_repository = "http://gitlab.egid.local/BioStats",
    mran = FALSE,
    targets = TRUE,
    python = TRUE
  )
}
}
