% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fex.R
\name{fex}
\alias{fex}
\title{Send file using fexsend}
\usage{
fex(..., zip_file = sprintf("\%s_archive.zip", Sys.Date()), internal = TRUE)
}
\arguments{
\item{...}{One or several path to files. A zip archive is created when several path are provided,}

\item{zip_file}{Name of the archive created when several path are provided.
Default `"20XX-XX-XX_archive.zip"`.}

\item{internal}{If internal fexsend binary should be used. Default `TRUE`.}
}
\value{
The URL of the upload file.
}
\description{
Documentation available at https://fex.belwue.de/fstools/fexsend.html
}
\examples{

if (interactive()) {
  fex("file1.txt", "file2.txt", internal = FALSE)
}

}
