% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report.R
\name{html_report}
\alias{html_report}
\title{html_report}
\usage{
html_report(
  toc_depth = 3,
  fig_width = 6.3,
  fig_height = 4.7,
  number_sections = TRUE,
  self_contained = TRUE,
  mathjax = "default",
  df_print = "kable",
  code_download = TRUE,
  ...
)
}
\arguments{
\item{toc_depth}{Depth of headers to include in table of contents}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{number_sections}{\code{TRUE} to number section headings}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method
uses a corresponding S3 method of \code{print}, typically
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition to the named methods you can
also pass an arbitrary function to be used for printing data frames. You
can disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}. See
\href{https://bookdown.org/yihui/rmarkdown/html-document.html#data-frame-printing}{Data
frame printing section} in bookdown book for examples.}

\item{code_download}{Embed the Rmd source code within the document and provide
a link that can be used by readers to download the code.}

\item{...}{Arguments to be passed to a specific output format function `rmarkdown::html_document`}
}
\value{
R Markdown output format to pass to render
}
\description{
These are simple wrappers of the output format functions like rmarkdown::html_document(),
and they added the capability of numbering figures/tables/equations/theorems
and cross-referencing them. See References for the syntax. Note you can also cross-reference
sections by their ID's using the same syntax when sections are numbered.
In case you want to enable cross reference in other formats, use markdown_document2
with base_format argument.
}
