% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioslides_presentation.R
\name{ioslides_presentation}
\alias{ioslides_presentation}
\title{ioslides_presentation}
\usage{
ioslides_presentation(
  smaller = FALSE,
  mathjax = "default",
  self_contained = TRUE,
  fig_width = 5.94,
  fig_height = 3.3,
  transition = 0.5,
  incremental = FALSE,
  ...
)
}
\arguments{
\item{smaller}{Use smaller text on all slides. You can also enable this for
individual slides by adding the \code{.smaller} attribute to the slide
header (see \emph{Presentation Size} below for details).}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{transition}{Speed of slide transitions. This can be "default",
"slower", "faster", or a numeric value with a number of seconds (e.g. 0.5).}

\item{incremental}{\code{TRUE} to render slide bullets incrementally.
Note that if you want to reverse the default incremental behavior for an
individual bullet you can preceded it with \code{>}.
For example: \emph{\code{> - Bullet Text}}.}

\item{...}{Arguments to be passed to a specific output format function `rmarkdown::ioslides_presentation`}
}
\value{
R Markdown output format to pass to render
}
\description{
Format for converting from R Markdown to an [ioslides](https://code.google.com/p/io-2012-slides/) presentation.
}
