% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerpoint_presentation.R
\name{powerpoint_presentation}
\alias{powerpoint_presentation}
\title{powerpoint_presentation}
\usage{
powerpoint_presentation(fig_width = 5.94, fig_height = 3.3, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{...}{Arguments to be passed to a specific output format function `rmarkdown::powerpoint_presentation`}
}
\value{
R Markdown output format to pass to render
}
\description{
Format for converting from R Markdown to a PowerPoint presentation. Pandoc v2.0.5 or above is required.
}
