% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_affiliation.R
\name{use_affiliation}
\alias{use_affiliation}
\title{Use UMR1283-8199 Affiliation}
\usage{
use_affiliation()
}
\value{
A character string with the current affiliation.
}
\description{
Use UMR1283-8199 Affiliation
}
