% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{use_python}
\alias{use_python}
\title{Use Python}
\usage{
use_python(project = ".", type = "virtualenv")
}
\arguments{
\item{project}{A character string. The path to a (RStudio) project.}

\item{type}{Type of Python environment to use, see \code{renv::use_python} for details.}
}
\description{
Setup directory structure for \code{python}.
}
