% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{Use targets package}
\usage{
use_targets(project = ".", ...)
}
\arguments{
\item{project}{A character string. The path to a project.}

\item{...}{Not used.}
}
\description{
Setup directory structure for \code{targets}.
}
