% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_xaringan.R
\name{use_xaringan}
\alias{use_xaringan}
\title{Use Xaringan Template}
\usage{
use_xaringan(
  path = "./scripts/slides",
  url = "https://github.com/umr1283/xaringan-template",
  overwrite = TRUE
)
}
\arguments{
\item{path}{A character string. The path to a directory within a RStudio project.}

\item{url}{A character string. A remote url, that ypically starts with \verb{https://github.com/} for public repositories,
and \verb{https://yourname@github.com/} or \verb{git@github.com/} for private repos.
You will be prompted for a password or pat when needed.}

\item{overwrite}{A logical. Should existing destination files be overwritten?}
}
\description{
Setup directory structure for \code{targets}.
}
