% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a project specific to UMR1283 storage architecture}
\usage{
create_project(
  path,
  analyst_name,
  git_repository,
  mran = FALSE,
  targets = TRUE,
  xaringan = FALSE,
  working_directory = NULL,
  ...
)
}
\arguments{
\item{path}{A character string. A path to where the project is to be created.}

\item{analyst_name}{A character string. The name of the analyst in charge of that project.}

\item{git_repository}{A character string. URL to the git server/repository.}

\item{mran}{A boolean. If \code{TRUE}, uses \code{paste0("https://mran.microsoft.com/snapshot/", Sys.Date())} as the CRAN repository.
Default is \code{FALSE}.}

\item{targets}{A boolean. If \code{TRUE}, uses \code{use_targets()} to create directory tree for use with \code{targets}.
Default is \code{TRUE}.}

\item{xaringan}{A boolean. If \code{TRUE}, uses \code{use_xaringan()} to create directory tree for use with \code{xaringan} template (\emph{i.e.}, "https://github.com/umr1283/xaringan-template").
Default is \code{FALSE}.}

\item{working_directory}{A character string. If specified, a symbolic link to the working directory will be created in the project directory under \code{outputs}, otherwise (default), \code{outputs} is a directory.}

\item{...}{not used}
}
\description{
Create a project specific to UMR1283 storage architecture
}
\examples{

if (interactive()) {
  create_project(
    path = "_test_project",
    analyst_name = "Analyst Name",
    git_repository = "http://gitlab.local/BioStats",
    mran = FALSE,
    targets = TRUE,
    xaringan = FALSE,
    working_directory = NULL
  )
}
}
