% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(
  path,
  analyst_name,
  working_directory,
  git_repository,
  mran = FALSE,
  targets = TRUE,
  python = FALSE,
  restart = interactive(),
  ...
)
}
\arguments{
\item{path}{A character string. A path to where the project is to be created.}

\item{analyst_name}{A character string. The name of the analyst in charge of that project.}

\item{working_directory}{A character string. A path to where outputs is to be generated.}

\item{git_repository}{A character string. URL to the git server/repository.}

\item{mran}{A boolean. If `TRUE`, uses `paste0("https://mran.microsoft.com/snapshot/", Sys.Date())` as the CRAN repository.
Default is `FALSE`.}

\item{targets}{A boolean. If `TRUE`, uses `use_targets()` to create directory tree for use with `targets`.
Default is `TRUE`.}

\item{python}{A boolean. If `TRUE`, uses `use_python()` to create `renv` directory tree for use with python.
Default is `FALSE`.}

\item{restart}{A boolean. If `TRUE`, restarts the RStudio session.}

\item{...}{not used}
}
\description{
create_project
}
\examples{

if (interactive()) {
  create_project(
    path = "/disks/PROJECT/_test_project",
    analyst_name = "Analyst Name",
    working_directory = "/disks/DATATMP",
    git_repository = "http://gitlab.egid.local/BioStats",
    mran = FALSE,
    targets = TRUE,
    python = TRUE
  )
}
}
