% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fex.R
\name{fex}
\alias{fex}
\title{Send file using fexsend}
\usage{
fex(..., zip_file = sprintf("\%s_archive.zip", Sys.Date()), internal = TRUE)
}
\arguments{
\item{...}{One or several paths to files. A zip archive is created when several paths are provided,}

\item{zip_file}{Name of the archive created when several paths are provided.
Default \code{"20XX-XX-XX_archive.zip"}.}

\item{internal}{If internal fexsend binary should be used. Default \code{TRUE}.}
}
\value{
The URL of the upload file.
}
\description{
Documentation available at https://fex.belwue.de/fstools/fexsend.html
}
\examples{

if (interactive()) {
  writeLines("file1", file.path(tempdir(), "file1.txt"))
  writeLines("file2", file.path(tempdir(), "file2.txt"))

  fex(file.path(tempdir(), "file1.txt")) # File sent as-is
  fex(file.path(tempdir(), c("file1.txt", "file2.txt"))) # Files sent as a ZIP archive
}

}
