% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_project.R
\name{migrate_project}
\alias{migrate_project}
\title{migrate_project}
\usage{
migrate_project(
  project = rprojroot::find_rstudio_root_file(),
  date,
  working_directory = "/disks/DATATMP",
  targets = FALSE,
  python = FALSE,
  restart = interactive(),
  ...
)
}
\arguments{
\item{project}{A character string. The path a RStudio project.}

\item{date}{A character string. The date of the MRAN to use, *e.g.*, 2021-01-25.}

\item{working_directory}{A character string. The root path to the working directory of the project.}

\item{targets}{A boolean. If `TRUE`, uses `use_targets()` to create directory tree for use with `targets`.
Default is `FALSE`.}

\item{python}{A boolean. If `TRUE`, uses `use_python()` to create `renv` directory tree for use with python.
Default is `FALSE`.}

\item{restart}{A boolean. If `TRUE`, restarts the RStudio session.}

\item{...}{not used}
}
\description{
migrate_project
}
