% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_project.R
\name{migrate_project}
\alias{migrate_project}
\title{\strong{CAUTION REQUIRED}: Migrate from an old version of project to renv setup}
\usage{
migrate_project(project = ".", date, working_directory, targets = FALSE, ...)
}
\arguments{
\item{project}{A character string. The path to a project.}

\item{date}{A character string. The date of the MRAN to use, \emph{e.g.}, 2021-01-25.}

\item{working_directory}{A character string. The root path to the working directory of the project.}

\item{targets}{A boolean. If \code{TRUE}, uses \code{use_targets()} to create directory tree for use with \code{targets}.
Default is \code{FALSE}.}

\item{...}{not used}
}
\description{
\strong{CAUTION REQUIRED}: Migrate from an old version of project to renv setup
}
