% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_project.R
\name{migrate_project}
\alias{migrate_project}
\title{migrate_project}
\usage{
migrate_project(path, date, working_directory = "/disks/DATATMP")
}
\arguments{
\item{path}{A character string. The path to the project to upgrade.}

\item{date}{A character string. The date of the MRAN to use, *e.g.*, 2021-01-25.}

\item{working_directory}{A character string. The root path to the working directory of the project.}
}
\description{
migrate_project
}
