% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{new_project}
\usage{
new_project(
  path,
  analyst_name = "Not specified",
  working_directory = "/disks/DATATMP",
  git_repository = "http://gitlab.egid.local/BioStat",
  ...
)
}
\arguments{
\item{path}{character}

\item{analyst_name}{character}

\item{working_directory}{character}

\item{git_repository}{character}

\item{...}{not used}
}
\value{
logical
}
\description{
new_project
}
