% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_affiliation.R
\name{use_affiliation}
\alias{use_affiliation}
\title{Inserm U1283 - CNRS UMR 8199 Affiliation}
\usage{
use_affiliation()
}
\value{
A character string with the current affiliation.
}
\description{
Inserm U1283 - CNRS UMR 8199 Affiliation
}
\examples{
use_affiliation()
}
