% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{use_python}
\alias{use_python}
\title{use_python}
\usage{
use_python(
  project = rprojroot::find_rstudio_root_file(),
  working_directory = "/disks/DATATMP",
  type = "virtualenv"
)
}
\arguments{
\item{project}{A character string. The path a RStudio project.}

\item{working_directory}{Path to the working temporary directory.}

\item{type}{Type of Python environment to use.}
}
\value{
results from `python -m ensurepip`.
}
\description{
use_python
}
