% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{Use targets package}
\usage{
use_targets(project = ".", working_directory = "/disks/DATATMP", ...)
}
\arguments{
\item{project}{A character string. The path a RStudio project.}

\item{working_directory}{Path to the working temporary directory.}

\item{...}{Not used.}
}
\description{
Setup directory structure for `targets`.
}
