% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{use_targets}
\usage{
use_targets(
  project = rprojroot::find_rstudio_root_file(),
  working_directory = "/disks/DATATMP",
  ...
)
}
\arguments{
\item{project}{A character string. The path a RStudio project.}

\item{working_directory}{Path to the working temporary directory.}

\item{...}{Any parameter to `targets::tar_script`.}
}
\description{
Setup directory structure for `targets`.
}
