% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_xaringan.R
\name{use_xaringan}
\alias{use_xaringan}
\title{Use Xaringan Template}
\usage{
use_xaringan(
  project = ".",
  sub_directory = "scripts",
  url = "https://github.com/umr1283/xaringan-template",
  overwrite
)
}
\arguments{
\item{project}{A character string. The path to a RStudio project.}

\item{sub_directory}{A character string. The path  to a directory in a RStudio project.}

\item{url}{remote url. Typically starts with \verb{https://github.com/} for public repositories,
and \verb{https://yourname@github.com/} or \verb{git@github.com/} for private repos.
You will be prompted for a password or pat when needed.}
}
\description{
Setup directory structure for \code{targets}.
}
